Implemented JPA Repository stubs Implemented tests for JPA and Hibernate Repositories

+++ sportsclub/sportsclub-domain/src/main/java/org/jboss/snowdrop/samples/sportsclub/domain/repository/ReservationRepository.java
@@ -9,5 +9,5 @@ public interface ReservationRepository extends Repository<Reservation, Long>
 {
    List<Reservation> getByCriteria(ReservationSearchCriteria criteria);
 
-   Integer countByCriteria(ReservationSearchCriteria criteria);
+   Long countByCriteria(ReservationSearchCriteria criteria);
 }
