+++ sportsclub/sportsclub-reservations-spring/src/main/java/org/jboss/snowdrop/samples/sportsclub/service/ReservationService.java
@@ -13,7 +13,7 @@ public interface ReservationService
 {
    List<Reservation> getReservations(Date fromDate, Date toDate, Integer nim, Integer max, List<EquipmentType> types);
 
-   Integer countReservationsForRange(Date fromDate, Date toDate, List<EquipmentType> types);
+   Long countReservationsForRange(Date fromDate, Date toDate, List<EquipmentType> types);
 
    void create(Reservation reservation);
 }
